Split base and generic handlers in AWS
