+++ cms/server/handlers/admin/__init__.py
@@ -23,7 +23,8 @@ from __future__ import unicode_literals
 
 from .base import \
     SimpleHandler, \
-    SimpleContestHandler, \
+    SimpleContestHandler
+from .main import \
     ResourcesHandler, \
     NotificationsHandler
 from .contest import \
